Fix Core Data type mappings for changeTypeRaw and offlinePasswordChangeCount
- changeTypeRaw: String → String? to match the pattern used by id,
  cipherId, and userId (Core Data KVC for string attributes returns
  NSString? which crashes on non-optional String)
- offlinePasswordChangeCount: Int → Int64 to match Core Data's
  documented Integer 64 mapping exactly
- Public interface stays Int with conversion at the managed object
  boundary

+++ BitwardenShared/Core/Vault/Services/OfflineSyncResolver.swift
@@ -57,7 +57,7 @@ actor DefaultOfflineSyncResolver: OfflineSyncResolver {
 
     /// The minimum number of offline password changes that triggers a server backup
     /// even when no conflict is detected (soft conflict threshold).
-    static let softConflictPasswordChangeThreshold: Int = 4
+    static let softConflictPasswordChangeThreshold: Int64 = 4
 
     // MARK: Properties
